refactor: rename and move a few things
